feat: Implement Cycle 01 - Core Framework & DFT Factory (Final)
- Added detailed architecture documentation in `mlip_autopipec/README.md`.
- Added dependency management guide in `DEPENDENCIES.md`.
- Updated `DFTResult` to use `Decimal` for precision energy storage.
- Refined `QERunner` error handling and cleanup logic.
- Enhanced UAT tests to verify input file generation content.
- Added missing docstrings to `mlip_autopipec/generator/alloy.py`.
- Verified all changes with strict linting, typing, and testing.

+++ mlip_autopipec/data_models/dft_models.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from enum import Enum
 from typing import Any
 
@@ -15,7 +16,7 @@ class DFTExceptionType(str, Enum):
 
 class DFTResult(BaseModel):
     uid: str
-    energy: float
+    energy: Decimal
     forces: list[list[float]] = Field(..., description="Nx3 array")
     stress: list[list[float]] = Field(..., description="3x3 array")
     succeeded: bool
